fix(codegen): DEPYLER-1140 Trait Bound Synthesis - Phase 1
Two E0277 fixes implemented:

1. Negative indexing usize fix (direct_rules_convert.rs:3041,3057)
   - Cast index to isize before negation: `(#index_expr) as isize`
   - Prevents E0277 "usize: Neg is not satisfied" when index is usize

2. From<Vec<T>> implementations for DepylerValue (rust_gen.rs:2789-2820)
   - Added From<Vec<String>>, From<Vec<i32>>, From<Vec<i64>>, From<Vec<f64>>,
     From<Vec<bool>>, From<Vec<&str>> for DepylerValue
   - Enables seamless typed vector to DepylerValue::List conversion
   - Fixes E0277 "DepylerValue: From<Vec<String>> not satisfied"

Also fixes missing `alias` field in Import struct for test files added in
DEPYLER-1136-1139.

+++ crates/depyler-core/src/direct_rules_convert.rs
@@ -3034,10 +3034,11 @@ impl<'a> ExprConverter<'a> {
                 // For negative literals, use runtime-safe indexing:
                 // base.get(base.len().wrapping_add(idx as usize)).cloned().unwrap()
                 // This handles -1 → len-1, -2 → len-2, etc.
+                // DEPYLER-1140: Cast to isize first to handle usize indices correctly
                 Ok(parse_quote! {
                     {
                         let _base = &#base_expr;
-                        let _idx = #index_expr;
+                        let _idx = (#index_expr) as isize;
                         let _actual_idx = if _idx < 0 {
                             _base.len().wrapping_sub((-_idx) as usize)
                         } else {
@@ -3049,10 +3050,11 @@ impl<'a> ExprConverter<'a> {
             } else {
                 // For non-negative or variable indices, generate simpler code
                 // but still handle potential negatives at runtime
+                // DEPYLER-1140: Cast to isize first to handle usize indices correctly
                 Ok(parse_quote! {
                     {
                         let _base = &#base_expr;
-                        let _idx = #index_expr;
+                        let _idx = (#index_expr) as isize;
                         if _idx < 0 {
                             _base[_base.len().wrapping_sub((-_idx) as usize)].clone()
                         } else {
